:fire: re-organized ingest directory, removed IngestPipelineRun
