+++ lamindb/db/__init__.py
@@ -44,22 +44,13 @@ A `SQLModel <https://sqlmodel.tiangolo.com>`__ session:
 
    session
 
-Ingest class:
-
-.. autosummary::
-   :toctree: .
-
-   Ingest
-   LinkIngest
-   LinkFeatureModel
-
 """
 
 from lndb_hub import hub  # noqa, currently not documented as being overhauled
 
 from ..dev.db import session  # noqa
+from . import ingest  # noqa
 from ._delete import delete  # noqa
-from ._ingest import Ingest, LinkIngest, ingest  # noqa
 from ._insert import insert  # noqa
 from ._link import LinkFeatureModel, link  # noqa
 from ._load import load  # noqa
