+++ lamindb/db/ingest/__init__.py
@@ -1,0 +1,39 @@
+"""Ingest data.
+
+Ingest is an operation that stores, tracks and annotates dobjects.
+
+Guide: :doc:`/db/guide/ingest`.
+
+Ingest operations:
+
+.. autosummary::
+    :toctree: .
+
+    add
+    remove
+    status
+    reset
+    commit
+
+Ingest helper classes:
+
+.. autosummary::
+    :toctree: .
+
+    Ingest
+    LinkIngest
+    LinkFeatureModel
+"""
+
+from .._link import LinkFeatureModel
+from ._ingest import (
+    Ingest,
+    LinkIngest,
+    add,
+    commit,
+    list_ingests,
+    print_logging_table,
+    remove,
+    reset,
+    status,
+)
